+++ package.json
@@ -17,9 +17,11 @@
     "@types/node": "^16.18.126",
     "@types/react": "^19.1.11",
     "@types/react-dom": "^19.1.8",
+    "@types/react-router-dom": "^5.3.3",
     "date-fns": "^4.1.0",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
+    "react-router-dom": "^7.8.2",
     "react-scripts": "5.0.1",
     "recharts": "^3.1.2",
     "typescript": "^4.9.5",
